Made main file for precomptopsort
